+++ docs/CHANGELOG.rst
@@ -118,7 +118,7 @@ Release changes
 
 First release. Provides,
 
-- Component architecture using python meta-classing, a plugin system using 
+- Component system using python meta-classing, a plugin system using 
   interface specifications and a wonderful configuration system.
 - Pluggable sub-commands accessible via `pa` script.
 - Web-framework to host more than one application in the same environment.